feat: moved some folders around
