+++ App.tsx
@@ -1,7 +1,7 @@
 import { config } from "@gluestack-ui/config";
 import { GluestackUIProvider, SafeAreaView } from "@gluestack-ui/themed";
 import { AuthenticationProvider } from "./src/contexts/useAuthenticationContext";
-import { AppNavigation } from "./src/components/modules/AppNavigation";
+import { AppNavigation } from "./src/navigation/AppNavigation";
 import { Platform, StatusBar } from "react-native";
 import ExpoStatusBar from "expo-status-bar/build/ExpoStatusBar";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
