--- src/components/modules/IllustrationStateLoading/index.tsx
@@ -1,5 +1,0 @@
-import { VStack } from "@gluestack-ui/themed";
-
-export const IllustrationStateLoading = () => {
-  return <VStack w="$full"></VStack>;
-};
