+++ src/components/modules/UserSymptomCard/index.tsx
