+++ src/components/organisms/SubHeader/index.tsx
@@ -23,12 +23,13 @@ export const SubHeader = ({
     return <SubHeaderSkeleton />;
   }
 
+  const formattedEntitesCount = currentEntries ? entriesCount : 0;
   const formattedPageCount = pageCount || 1;
 
   return (
     <HStack width="$full" justifyContent="space-between" height="$10">
       <Text.Small color="sky_blue">
-        Displaying {currentEntries} of {entriesCount} {label}
+        Displaying {currentEntries} of {formattedEntitesCount} {label}
       </Text.Small>
       <Text.Small color="sky_blue">
         Page {currentPage} of {formattedPageCount}
