+++ src/navigation/AppNavigation/index.tsx
@@ -1,7 +1,7 @@
 import { NavigationContainer } from "@react-navigation/native";
-import { PrivateNavigation } from "../../../navigation/PrivateNavigation";
-import { PublicNavigation } from "../../../navigation/PublicNavigation";
-import { useAuthenticationContext } from "../../../contexts/useAuthenticationContext";
+import { PrivateNavigation } from "../PrivateNavigation";
+import { PublicNavigation } from "../PublicNavigation";
+import { useAuthenticationContext } from "../../contexts/useAuthenticationContext";
 
 export const AppNavigation = () => {
   const { state } = useAuthenticationContext();
